Adding initial role
Supports creating groups, with an option GID and system/non-system flag, and removing groups.

This is a very straight forward role so there isn't much in the way of documentation (as I assume users know what an OS group is, this is just a way to implement them).

This role is designed to be compatible with the system-users role where secondary groups are needed (as the system-users role doesn't create such groups).